Fix possible merge error in package-lock.json
Hi Jan from Depfu here. I noticed an error in our update process and it looks like some process or a manual merge error (I did not investigate) removed a crucial comma from the package-lock.json which throws a wrench into our gears.

This fixes the JSON validation error (so that the JSON can be parsed again) but I am not sure if this makes the package-lock.json actually valid for npm.